// DXC_Frontend/src/api/models/restaurantDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { RestaurantImageDto } from './restaurantImageDto'

export interface RestaurantDto {
  publicId?: string
  /** @nullable */
  name: string | null
  /** @nullable */
  description?: string | null
  /** @nullable */
  address?: string | null
  /** @nullable */
  phoneNumber?: string | null
  /** @nullable */
  operatingHours?: string | null
  /** @nullable */
  schedule?: string | null
  /** @nullable */
  latitude?: number | null
  /** @nullable */
  longitude?: number | null
  /** @nullable */
  vR360Link?: string | null
  /** @nullable */
  category?: string | null
  /** @nullable */
  averagePriceRange?: string | null
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  /** @nullable */
  images?: RestaurantImageDto[] | null
}